\name{xergm-package}
\docType{package}
\alias{xergm-package}
\title{Extensions of Exponential Random Graph Models (ERGM)}
\description{Extensions of Exponential Random Graph Models (ERGM).}
\details{
The \pkg{xergm} package implements extensions of exponential random 
graph models, in particular Temporal ERGMs (\pkg{btergm}),  
Generalized ERGMs (\pkg{gergm}), Temporal Network Autocorrelation 
Models (\pkg{tnam}), and Relational Event Models. This package acts 
as a meta-package for the packages \pkg{btergm}, \pkg{gergm}, and 
\pkg{tnam}. To display citation information, type 
\code{citation("xergm")}.
}
\author{
Philip Leifeld (\url{http://www.philipleifeld.de})

Skyler J. Cranmer (\url{http://www.skylercranmer.net})

Bruce A. Desmarais (\url{http://polisci.la.psu.edu/people/bruce-desmarais/})
}
